# Conflicts:
#	src/event/event.service.ts
#	src/event/notifications/notifications.service.ts

+++ src/event/event.service.ts
@@ -7,6 +7,9 @@ import { Event } from '../model/event/event.entity';
 import { EventBody } from './interfaces/event-body.interface';
 import { EventCategory } from '../model/event/event-category.enum';
 import { EventAttendee } from '../model/event-attendee/event-attendee.entity';
+import { NotificationsService } from './notifications/notifications.service';
+
+
 import * as moment from 'moment';
 import * as _ from 'lodash';
 
@@ -14,6 +17,7 @@ import * as _ from 'lodash';
 export class EventService {
 
   constructor(
+    private readonly notificationsService: NotificationsService,
     @InjectRepository(EventRepository)
     private readonly eventRepository: EventRepository,
     @InjectRepository(EventAttendee)
@@ -44,6 +48,21 @@ export class EventService {
   }
 
   async deleteOwnEvent(user, id) {
+    const allEvents: Event[] = await this.allEvents();
+    let filteredEvents: Event[] = allEvents;
+    let eventofuser = filteredEvents.filter(event => event.id == id);
+    let userevent = eventofuser.map(ea => ea.user).pop();
+    let event = await this.oneEvent(id);
+    console.log(userevent);
+    console.log(event);
+    if (userevent.id != user.id) throw new BadRequestException('You can not delete an event you are not the owner');
+    let AttendingListDeletedEvent = await this.getEventAttendingList(event);
+    let userToBeNotified: User [] = AttendingListDeletedEvent.map(ea => ea.user);
+    console.log(userToBeNotified);
+    let body: string = 'La actividad ' + event.name + ' ha sido borrada';
+    console.log(body);
+    await this.notificationsService.addNotification(userToBeNotified, body);
+    AttendingListDeletedEvent.forEach(ea => this.eventAttendeeRepository.delete({ id: ea.id }));
     return await this.eventRepository.delete({ id });
   }
 
@@ -96,6 +115,13 @@ export class EventService {
     if (oldAttendees.length > 0) event.attendance = (newAttendees.length / oldAttendees.length);
     else event.attendance = 0;
     await this.eventRepository.save(event);
+
+    //Notifications
+    let UserToBeNotified: User [] = newAttendees.map(ea => ea.user); //FIXME
+    console.log(UserToBeNotified);
+    let body: string = '¡La actividad  ' + event.name + ' ya la puede valorar!';
+    console.log(body);
+    this.notificationsService.addNotification(UserToBeNotified, body);
 
     if (newAttendees.length != 0) {
       for (const attendee of attendees) {
